Change gsl, glm

+++ AmethystAPI/include/glm/gtx/scalar_multiplication.hpp
@@ -1,9 +1,6 @@
-/// @ref gtx_scalar_multiplication
+/// @ref gtx
 /// @file glm/gtx/scalar_multiplication.hpp
 /// @author Joshua Moerman
-///
-/// @defgroup gtx_scalar_multiplication GLM_GTX_scalar_multiplication
-/// @ingroup gtx
 ///
 /// Include <glm/gtx/scalar_multiplication.hpp> to use the features of this extension.
 ///
@@ -35,9 +32,6 @@
 
 namespace glm
 {
-	/// @addtogroup gtx_scalar_multiplication
-	/// @{
-
 	template<typename T, typename Vec>
 	using return_type_scalar_multiplication = typename std::enable_if<
 		!std::is_same<T, float>::value       // T may not be a float
@@ -60,7 +54,7 @@ namespace glm
 	template<typename T> \
 	return_type_scalar_multiplication<T, Vec> \
 	operator/(Vec lh, T const& s){ \
-		return lh *= 1.0f / static_cast<float>(s); \
+		return lh *= 1.0f / s; \
 	}
 
 GLM_IMPLEMENT_SCAL_MULT(vec2)
@@ -78,5 +72,4 @@ GLM_IMPLEMENT_SCAL_MULT(mat4x3)
 GLM_IMPLEMENT_SCAL_MULT(mat4)
 
 #undef GLM_IMPLEMENT_SCAL_MULT
-	/// @}
 } // namespace glm
